Remove "srb2" namespace requirement and remove uneeded code.
This should remove the need to have srb2 as your namespace to access extra features UZB adds for SRB2. This is handy if you are making an SRB2 source code derived project and you want to use a custom namespace in engine.

This also removes the code for the old UDMF menus since they will no longer be accessible. Since this is a SRB2 editor, I'm not really sure people are going to map for other DOOM sourceports in UZB regardless...